File Structure: Rename sequencer --> video_editing
I did not want to make this change now but the reasoning is, it is better to make the changes now rather than later.

Currently URL look like: https://docs.blender.org/manual/en/dev/sequencer/sequencer/index.html

Which are a bit strange.

This change will allow us to expand this chapter more in the future.

git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk@5558 c4de1f47-6596-e411-a384-0024e86c2797