Revert the addition of sphinx-panel (7176 & 7177)
This added a new dependency for building without review,
Preventing a default sphinx installation from being usable to build the docs.

Hiding sections of a document has down sides too -
that text can't be searched on the page.

If a short list of error messages in one place is better we could look into
ways of doing this with the current system.

See replies to https://developer.blender.org/rBM7176



git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk@7194 c4de1f47-6596-e411-a384-0024e86c2797