A note on Eyedropper consistancy
The default option for view transform (Properties -> Render Properties (any engine) -> Color Management -> View Transform) is set to Filmic. This can lead to eyedropper returning a color darker than what the sample actually is. This is especially evident when we use the eyedropper to sample the color of an object and then set the color of the same object to the sampled color. Normally you expect the color to remain the same when doing so but with the Filmic option, every time you do this you get a darker color until you reach black.

I decided to add this edit to the manual since this can be really frustrating to people that are relatively new to blender like myself and that I couldn't find anything about this issue by searching for issues related to eyedropper. Hopefully this would save someone's time.

Author: @WetBird

Differential Revision: https://developer.blender.org/D11884


git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk@8229 c4de1f47-6596-e411-a384-0024e86c2797