Config: Disable copy sources
This is no longer needed now that we link to a only copy of the sources.

In previous versions of sphinx this was still needed to control search
summeries but that is no longer needed after the recent update.

Note however sphinx 4.1.2 reverted this which breaks search summaries.
sphinx 4.2 should fix this so we will be skipping the rest of 4.1.x versions.

See: https://github.com/sphinx-doc/sphinx/issues/9456


git-svn-id: https://svn.blender.org/svnroot/bf-manual/trunk@8334 c4de1f47-6596-e411-a384-0024e86c2797